add pavics-component config files
Assume `thunderbird` repo is checkout along-side with `birdhouse-deploy`
repo because we use `../../thunderbird/pavics-component` path for volume
mount.

`env.local` config to add to activate this component:
```
  export EXTRA_CONF_DIRS="$EXTRA_CONF_DIRS /somepath/thunderbird/pavics-component"
  export AUTODEPLOY_EXTRA_REPOS="$AUTODEPLOY_EXTRA_REPOS /somepath/thunderbird/pavics-component"
  export THUNDERBIRD_IMAGE="tlvu/thunderbird:i3-gen-climos"
```

The bird will be accessible at
https://$PAVICS_FQDN/twitcher/ows/proxy/thunderbird/wps or
http://$PAVICS_FQDN:8099/wps.

Public port 8099, same as the existing `docker-compose.yml`.

Magpie permissions give full public access to the bird.  This should be
adjusted later.  Or this file can simply be replaced for a different
permission level.